+++ CHANGELOG.md
@@ -9,9 +9,10 @@ This changelog track changes to the qoqo project starting at version 0.5.0
 ### Added
 
 * Semver-style version checking for Circuit serialization. In beta mode (0.y.z) minor version must match (y_library == y_data) in release mode (x.y.z) major version must match (x_library == x_data) and minor version of library must exceed minor version of data (y_library >= y_data).
-* `json_schema` implementing `JsonSchema` from schemars for roqoqo data structures
+* `json_schema` implementing `JsonSchema` from schemars for roqoqo data structures.
 * `roqoqo-test` extended by two new functions for stochastic gate tests: `construct_random_circuit` and `add_random_multi_qubit_gate`. 
 * A conversion function `to_single_qubit_gate` has been implemented for the OperateSingleQubitGate trait.
+* The multiplication function `mul` added to the python interface, i.e. qoqo, for Single Qubit Gates.
 
 ### Changed